user validator, Usermanager 코드 분리
